requirements.txt
googletrans-py
ollama